fix: add .toDouble()/.toInt() after .clamp() for Dart 3.10+ type safety
Dart 3.10+ returns 'num' from .clamp() instead of 'double'/'int',
causing compile errors when assigned to typed variables.

Fixed 18 files:
- stay_balance_calculator.dart (2 errors reported by CI)
- booking_price_adjustment_service.dart
- booking_derived_fields_service.dart (2 locations)
- price_adjustment_service.dart
- gemini_service.dart (3 locations)
- booking_computed_stream_service.dart (4 locations)
- enhanced_booking_calculation_service.dart (3 locations)
- booking_payment_screen.dart (5 locations)
- booking_checkout_screen.dart (3 locations)
- bookings_list.dart
- debts_list.dart
- create_debt_from_booking.dart
- guest_payments_detail_report_screen.dart
- active_bookings_reminder_screen.dart
- data_protection_screen.dart
- data_usage_manager.dart
- guest_payment_calculation_service.dart
- shimmer_loading.dart

+++ mobile/lib/screens/debts/debts_list.dart
@@ -857,7 +857,7 @@ class _DebtsListScreenState extends ConsumerState<DebtsListScreen>
     void recalculate() {
       final total = CurrencyFormatter.parseAmount(totalCtrl.text) ?? 0;
       final paid = CurrencyFormatter.parseAmount(paidCtrl.text) ?? 0;
-      final remaining = (total - paid).clamp(0.0, double.infinity);
+      final remaining = (total - paid).clamp(0.0, double.infinity).toDouble();
       remainingCtrl.text = CurrencyFormatter.formatAmount(remaining);
     }
