+++ mobile/lib/services/enhanced_booking_calculation_service.dart
@@ -250,7 +250,7 @@ class EnhancedBookingCalculationService {
         }
       }
 
-      final int finalRate = (baseRate + adjustmentTotal).clamp(0, baseRate * 100);
+      final int finalRate = (baseRate + adjustmentTotal).clamp(0, baseRate * 100).toInt();
 
       breakdown.add(
         NightlyBreakdown(
@@ -306,12 +306,12 @@ class EnhancedBookingCalculationService {
     final totalDiscount =
         booking.discountType == 'total' ? _asInt(booking.discount) : 0;
     if (totalDiscount > 0) {
-      totalDue = (totalDue - totalDiscount).clamp(0, totalDue);
+      totalDue = (totalDue - totalDiscount).clamp(0, totalDue).toInt();
       totalAdjustments -= totalDiscount;
     }
 
     final totalPaid = await _getTotalPayments(booking);
-    final remaining = (totalDue - totalPaid).clamp(0, totalDue);
+    final remaining = (totalDue - totalPaid).clamp(0, totalDue).toInt();
 
     return FinancialSummary(
       subtotal: subtotal,
